Minor update to ensure j1 remains integer.
git-svn-id: https://scm.physics.ox.ac.uk/svn/radtrancode/trunk@552 4dc0e162-f54d-4b9e-8a18-81c7a7f2ddce

+++ radtran/cirsradg/cirsradg_wave.f
@@ -844,7 +844,7 @@ C           matrix inversion crashing
 
 
 C           Check to see if this is a limb path
-            j1=0.5*nlays
+            j1=int(0.5*nlays)
 
             p1=press(layinc(j1,ipath))
             p2=press(layinc(nlays,ipath))
